software/requirements.txt
flask
paho-mqtt
python-telegram-bot==20.4
requests
pytest
